+++ eslint.config.mjs
@@ -3,7 +3,7 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import js from '@eslint/js';
 import { FlatCompat } from '@eslint/eslintrc';
-import jest from 'eslint-plugin-jest';
+import vitest from 'eslint-plugin-vitest';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -26,7 +26,7 @@ export default [
 			'plugin:@typescript-eslint/eslint-recommended',
 			'plugin:@typescript-eslint/recommended'
 		),
-		jest.configs.recommended
+		vitest.configs.recommended
 	),
 	{
 		settings: {